chore: organize NextStepCards
