+++ components/dashboard/NextStepCard.tsx
