+++ components/dashboard/OnboardingCohortRegister.tsx
